pyproject.toml
[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false

[project]
name = "python-template"
description = ""
requires-python = ">=3.12"
dependencies = [
    "pandas (>=2.2.3,<3.0.0)",
    "rich (>=13.9.4,<14.0.0)",
    "tqdm (>=4.67.1,<5.0.0)",
    "faker (>=36.1.1,<37.0.0)",
    "go-task-bin (>=3.41.0,<4.0.0)",
    "gitpython (>=3.1.44,<4.0.0)",
    "black (>=25.1.0,<26.0.0)",
    "arrow (>=1.3.0,<2.0.0)",
    "pylint (>=3.3.4,<4.0.0)",
    "vulture (>=2.14,<3.0)",
    "mkdocstrings-python (>=1.16.1,<2.0.0)",
    "mkdocs-material (>=9.6.4,<10.0.0)",
]
